+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 ### Added 
 - `MainActivity`, `MainViewModel`
+- `MapFragment`, `StationCluster`